# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notify Center",
        "description": "通知中心 api description",
        "contact": {
            "email": "developer@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/staff/api",
            "description": "測試區"
        },
        {
            "url": "http://localhost/XXXXX",
            "description": "Localhost"
        }
    ],
    "paths": {
        "/": {},
        "/notify/{service}/{type?}": {
            "post": {
                "tags": [
                    "Notify"
                ],
                "summary": "新增通知訊息",
                "description": "\n    Parameters 參數文件\n\n    service:\n     * 參數類型: string\n     * 參數說明: 通知服務類型\n     * 是否必傳: Y\n     * 注意事項: {service} 服務類型可選: gmail, line, jandi, slack\n\n    type:\n     * 參數類型: string\n     * 參數說明: 通知情境類型\n     * 是否必傳: Y\n     * 注意事項: 每個通知服務開放情境如下\n    - gmail: exchange_rate(台灣銀行平均匯率通知), social_media_case(最新社群案例通知), resign(員工離退通知), invoice(電子發票開立通知)\n    - line:  尚未開放\n    - jandi: 尚未開放\n    - slack: 尚未開放\n\n    recipient_name:\n     * 參數類型: string\n     * 參數說明: 接收者姓名\n     * 是否必填: Y\n\n    email:\n     * 參數類型: string\n     * 參數說明: 電子信箱\n     * 是否必填: N\n\n    carbon_copy:\n     * 參數類型: array\n     * 參數說明: 副本\n     * 是否必填: N\n\n    blind_carbon_copy:\n     * 參數類型: array\n     * 參數說明: 副本\n     * 是否必填: N\n\n    subject:\n     * 參數類型: string\n     * 參數說明: 通知訊息主旨\n     * 是否必填: Y\n\n    content:\n     * 參數類型: Json\n     * 參數說明: 通知訊息內文\n     * 是否必填: Y\n     * 注意事項: 每種通知情境所需內文格式不同\n\n    attachment:\n     * 參數類型: array\n     * 參數說明: 附件\n     * 是否必填: N\n\n---------------------------------------------------\n\n    每個通知情境所需內文(content)如下\n\n    exchange_rate:\n     * 範例: {'year':'2024','month':'02'}\n     * 參數:     year       month\n     * 說明:   匯率表年份   匯率表月份\n     * 類型:    string     string\n\n    social_media_case:\n     * 範例: {'month':'3','cases':['案例標題',...]}\n     * 參數:    month            cases\n     * 說明: 案例分享當前月份    社群案例標題\n     * 類型:   string         array[string]\n\n    invoice:\n     * 範例: {'number':'ZA17444891','time':'2024/03/05'}\n     * 參數:   number          time\n     * 說明:   發票號碼      發票開立時間\n     * 類型:   string         string\n\n    resign:\n     * 範例: {'resignations':[{'employee_id':'','name':'','name_en':'','department':'','resignation_date':'','last_working_day':'','note':''},...]}\n     * 參數:   employee_id     name      name_en     department     resignation_date    last_working_day    note\n     * 說明:     員工編號      員工姓名   員工英文名字    員工所屬部門          離職日期            最後工作日        備註\n     * 類型:     string       string     string        string             string             string        string\n    ",
                "operationId": "f339897e545704aa1f97b5578c4c11a2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ]
                            },
                            "example": {
                                "recipient_name": "張育誠",
                                "email": "theyouchman@gmail.com",
                                "carbon_copy": [
                                    "johnny31258@gmail",
                                    "johnny.chang@js-adways.com.tw"
                                ],
                                "blind_carbon_copy": [
                                    "johnny31258@gmail",
                                    "johnny.chang@js-adways.com.tw"
                                ],
                                "subject": "台灣銀行2024年02月份平均匯率表",
                                "content": {
                                    "year": "2024",
                                    "month": "02"
                                },
                                "attachment": [
                                    "2024ExchangeRate-每月一號提供.xlsx"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "請求成功",
                        "headers": {
                            "test": {
                                "description": "Employee request ID",
                                "schema": {
                                    "type": "integer",
                                    "format": "int64"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "通知訊息創建成功",
                                    "details": {
                                        "recipient": "張育誠",
                                        "email": "theyouchman@gmail.com",
                                        "subject": "台灣銀行2024年02月份平均匯率表",
                                        "service": "Gmail",
                                        "type": "exchange_rate",
                                        "sent_at": "2024-03-08 12:16:15"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/401"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        }
    },
    "components": {
        "responses": {
            "200": {
                "description": "成功",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "status": 200,
                            "message": "OK"
                        }
                    }
                }
            },
            "400": {
                "description": "客戶端錯誤",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "status": 400,
                            "message": "Bad Request"
                        }
                    }
                }
            },
            "401": {
                "description": "身份驗證失敗",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "status": 401,
                            "message": "Unauthorized"
                        }
                    }
                }
            },
            "404": {
                "description": "找不到請求的資源",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "status": 404,
                            "message": "Not Found"
                        }
                    }
                }
            },
            "405": {
                "description": "不支援此方法",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "status": 405,
                            "message": "Method Not Allowed"
                        }
                    }
                }
            },
            "500": {
                "description": "伺服器發生錯誤",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "status": 500,
                            "message": "伺服器發生錯誤"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}